cli/requirements-cli.txt
oauth2client
google-api-python-client
google_auth_oauthlib
httplib2shim
#progress
#progressbar2